Refactors audit endpoint to use query params
Changes the /audit endpoint to accept the URL as a query parameter
instead of in the request body. This allows for GET requests to
trigger audits. Also extracts the audit execution logic into a
separate function to reduce code duplication.

Additionally, modifies the lighthouse execution to not delete the
temporary report file to permit debugging.
Adds docker-compose network to allow VPN support.